pyproject.toml
[project]
name = "hsk-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mff-pretty-cli>=0.2.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
]
